requirements.txt
requests>=2.31.0
feedparser>=6.0.10
python-dotenv>=1.0.0
openai>=1.0.0
praw>=7.7.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
colorama>=0.4.6
schedule>=1.2.0
aiohttp>=3.8.5
asyncio-throttle>=1.0.2
vaderSentiment>=3.3.2
python-dateutil>=2.8.2
pytz>=2023.3
